requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6
nltk>=3.8.0
scikit-learn>=1.3.0
numpy
python-Levenshtein>=0.20.0
rank-bm25>=0.2.0
datasets>=2.14.0
